moved most common functions to common.utils. also added hash_file func
